Allow a third Postgres server to join a group.
This makes it an extra secondary server, and for that we transition from
PRIMARY to the new state JOIN_PRIMARY. This means we still have a secondary
to failover to should the need appear, and we are concurrently adding a
second standby server to our group.

This is all still very much Work In Progress, as I don't expect anything
other than registration at the right timing to kind of work.

We still have to fix our usage of the replication slot name from the
primary, where it doesn't make sense anymore. Also we have to create
replication slots on every other secondary now. And so much more.