Restrict standby registration to one-at-a-time.
Given our current FSM implementation and protocol it's impossible to know
which standby is free to reach CATCHING UP when the primary has prepared the
replication settings (HBA and replication slot). So rather than figuring out
a way to extend the protocol to accept registering several standby at a
time, we implement a retryable error on the monitor.

And we also implement a retry-loop on the client side. Which is implemented
as a tail-call recursive call at the moment, because that's easy.