Return the LSN and isPrimary? in pgautofailover.get_other_nodes.
We will need this information to be able to create and manager replication
slots on each standby. When using multiple standbys we want to make sure we
keep the WAL we got for any other standby, in case we lose the primary. To
prepare for being able to do that, return the information we need from the
call to get_other_nodes.

In passing, add a JSON output to the `pg_autoctl do show others` command.
That makes it easy to programmatically parse the output, if needed. And then
we can display a human oriented table output for the command when --json is
not used.